custom-elements/concepts/custom-elements-type-allowed-chars-first-char.html

<!DOCTYPE html>
<html>
<head>
<title>First char allowed for custom element type </title>
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type identifies a custom element interface and is a sequence of characters that must match the NCName production">
<link rel="help" href="https://dvcs.w3.org/hg/webcomponents/raw-file/default/spec/custom/index.html#concepts">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function testArray(doc, array, positive) {
    for (var i=0; i<array.length; i++){
        var c = array[i];
        var name = String.fromCharCode(c) + '-x' + c.toString(16);
        if (positive) {
            try {
                doc.register(name);
            } catch (e) {
                assert_unreached('The custom element name \'' + name +
                       '\' (char code U+' + c.toString(16) + ') ' +
                       'should be registered without errors');
            }
        } else {
            assert_throws(null,
                          function(){doc.register(name);},
                          'Registering invalid custom element name \'' + name +
                            '\' (char code U+' + c.toString(16) + ') ' +
                            'should fail');
        }
    }
}

function testRangesArray(doc, array, positive) {
    for (var i=0; i<array.length; i+=2){
        var low = array[i];
        var high = array[i+1];
        for (var c=low; c<=high; c++){
            var name = String.fromCharCode(c) + '-x' + c.toString(16);
            if (positive){
                try {
                    doc.register(name);
                } catch (e) {
                    assert_unreached('The custom element name \'' + name +
                       '\' (char code U+' + c.toString(16) + ') ' +
                       'should be registered without errors');
                }
            } else {
                assert_throws(null,
                                function(){doc.register(name);},
                              'Registering invalid custom element name \'' + name +
                                 '\' (char code U+' + c.toString(16) + ') ' +
                                 'should fail');
            }
        }
    }
}

test(function(){
    testArray(newHTMLDocument(), ['_'], true);
}, 'Registering valid custom element name starting with \'_\' char');

test(function(){
    testArray(newHTMLDocument(), BASE_CHARS_SINGLE, true);
}, 'Registering valid custom element name starting with single base chars');

test(function(){
    testRangesArray(newHTMLDocument(), BASE_CHARS_RANGES, true);
}, 'Registering valid custom element name starting with base chars ranges');

test(function(){
    testArray(newHTMLDocument(), IDEOGRAPHIC_CHARS_SINGLE, true);
}, 'Registering valid custom element name starting with single ideographic chars');

test(function(){
    testRangesArray(newHTMLDocument(), IDEOGRAPHIC_CHARS_RANGES, true);
}, 'Registering valid custom element name starting with ideographic chars ranges');

test(function(){
    testArray(newHTMLDocument(), ['.', '-', ':', '+', '='], false);
}, 'Registering custom element starting with invalid character (general) should fail');

test(function(){
    testArray(newHTMLDocument(), COMBINING_CHARS, false);
}, 'Registering custom element starting with invalid character (combining char) should fail');

test(function(){
    testArray(newHTMLDocument(), EXTENDER_CHARS, false);
}, 'Registering custom element starting with invalid character (extender char) should fail');

test(function(){
    testRangesArray(newHTMLDocument(), DIGIT_CHARS_RANGES, false);
}, 'Registering custom element starting with invalid character (digit char) should fail');


</script>
</body>
</html>
